Room reservations w/ same room, different guests option
Again, needs some final debugging, so i'm not sure if we can finish on time to use this version.

I thought Room shouldn't have an array of guests so i remove that and added the array of guests in Guest. In that case a guest can bring many guests with him but a room can't be filled with guests. Sounds confusing but is easier if you see it in the code.

1. isEmpty, checks if there is a guest attached to the room object
2. isReserved, checks if room is reserved by comparing date objects
3. setResDates, gets reservation dates from user, creates and store dates if room is not reserved. Also stores (by calling set method) the guest's reservation dates. (That final part was not the best way to do it)
4. resDates, arrayList of all guests' reservation dates
5. displayDates, displays from arrayList of dates